% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgez.R
\name{load_data}
\alias{load_data}
\title{Load Data}
\usage{
load_data(file, inputs = "auto", output, inputs.class = "auto",
  output.class = "auto", train.size = 1, seed = 1,
  na.handle = "inf", max.levels = 50)
}
\arguments{
\item{file}{\strong{Character}. The link to the file containing the data.
The data are imported with the \code{fread} function from
the \code{data.table} package (\link[data.table]{fread}), so the format
must be consistent with a csv file.}

\item{inputs}{\strong{Character vector}. Vector of the column names for
the inputs of the model. Only those columns will be used for the model.
Using the "auto" value will use as inputs all the columns from the
table except the one labelled as output.}

\item{output}{\strong{Character}. A single string specifying the name
of the output column for the model training.}

\item{inputs.class}{\strong{Character vector}. A vector specifying the
classes for the input column. If set to "auto", the classes will be
determined from the output of the \link[data.table]{fread} function.
Else, it must me a vector whose size is exactly the number of input and
whose values can only be \emph{num} (for numerical inputs) and \emph{cat}
 (for categorical inputs).}

\item{output.class}{\strong{Character}. Class for output. If set to "auto",
the class will be determide from the output of the \link[data.table]{fread}
function. Else, it must be equal to \emph{num} or \emph{cat} for numerical
or categorical inputs.}

\item{train.size}{\strong{Numeric}. Size for training set for the future
model. Can go from 0 (no training set: will produce an error) to 1 (no test
set).}

\item{seed}{\strong{Numeric}. Seed for reproducibility of the results.}

\item{na.handle}{\strong{Character}. Way to handle na value in numeric
inputs. Five possibilities have been implemented:
\itemize{
 \item{\emph{inf}}: replace missing values with \code{Inf}.
 \item{\emph{mean}}: replace missing values with the mean of the column.
 \item{\emph{median}}: replace missing values with the median of the column.
 \item{\emph{max}}: replace missing values with the max of the column.
 \item{\emph{min}}: replace missing values with the min of the column.
}}

\item{max.levels}{\strong{Numeric}. Maximum number of levels admitted for
a category. This parameters is here to make sure that the model does not
have to many input data when transformed into a one-hot encoded matrix.}
}
\value{
A list with following values:
\itemize{
 \item{\strong{train}}: training set for the model, with a matrix for the
 input values and a vector for the target variables.
 \item{\strong{test}}: test set for the model, on the same format that the
 training set
 \item{\strong{formula}}: the formula used for constructing the model matrix
 and that is applied when running the model.
 \item{\strong{template}}: an empty \code{data.table} that has saved all the
 input values and that is used to appropriately format data when using
 the prediction function.
 \item{\strong{data}}: A data.table with the cleaned data and an additional
 logical column, \emph{train}, that indicates which data are used in the
 training data set.
}
}
\description{
\code{load_data} returns a list with all the prepared element for loading
and preparing the data for xgboost modelling. The model principally relies
on two categories of input: numeric (\emph{num}) and category (\emph{cat}).
}
\examples{
temp.file <- tempfile()
data(iris)
fwrite(iris, temp.file, row.names = F)
d <- load_data(temp.file, output = "Species")

}
