% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgez.R
\name{xg_auto_ml}
\alias{xg_auto_ml}
\title{Auto Machine Learning}
\usage{
xg_auto_ml(data, param)
}
\arguments{
\item{data}{\strong{Character}. The link to the file containing the data.
The data are imported with the \code{fread} function from
the \code{data.table} package (\link[data.table]{fread}), so the format
must be consistent with a csv file.}

\item{param}{\strong{Character}. A link to a YAML file with all the
the parameters for calibrating the model. See
system.file("extdata", "ex_param.yml", package = "ezXg") for an example.}
}
\value{
A calibrated model directly usable for prediction, consistent
with a model produced by \link[ezXg]{xg_train}.
}
\description{
\code{xg_auto_ml} compiles all the functions for calibrating a xgboost
model into a single one.
}
\examples{
md <- xg_auto_ml(system.file("extdata", "titanic.csv", package = "ezXg"),
                 system.file("extdata", "ex_param.yml", package = "ezXg"))

}
