% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgez.R
\name{xg_bo}
\alias{xg_bo}
\title{Bayesian Optimization}
\usage{
xg_bo(data, eta = c(0.05, 0.3), gamma = c(0, 0.5), max_depth = c(1L,
  15L), colsample_bytree = c(0.5, 1), min_child_weight = c(1L, 7L),
  nrounds = 100, nthread = 2, cv = 5, seed = 1,
  objective = "auto", init_grid_dt = NULL, kernel = list(type =
  "exponential", power = 2), init_points = 10, n_iter = 50,
  acq = "ucb", kappa = 2.576, eps = 0, verbose = TRUE)
}
\arguments{
\item{data}{\strong{Object}. A data structure created by the call of the
\link[ezXg]{xg_load_data} function.}

\item{eta}{\strong{Numeric Vectors}. Eta parameter min and max bounds.}

\item{gamma}{\strong{Numeric Vector}. Gamma parameter min and max bounds.}

\item{max_depth}{\strong{Numeric Vector}. Max_depth parameter min and max bounds.
Use "L" suffix to indicate integer hyperparameter.}

\item{colsample_bytree}{\strong{Numeric Vector}. Colsample_bytree parameter
min and max bounds.}

\item{min_child_weight}{\strong{Numeric Vector}. Min_child_weight parameter
min and max bounds. Use "L" suffix to indicate integer hyperparameter.}

\item{nrounds}{\strong{Numeric}. Nrounds parameter for xgboost calibration.
See \link[xgboost]{xgb.train} for more details.}

\item{nthread}{\strong{Numeric}. Nthread parameter for xgboost calibration.
See \link[xgboost]{xgb.train} for more details.}

\item{cv}{\strong{Numeric}. Number of folds in cross validation. Needs
to be more than 2.}

\item{seed}{\strong{Numeric}. Seed for computation reproducability.}

\item{objective}{\strong{Character}. Objective function for the
optimization. . Eta parameter for xgboost calibration. See
 \link[xgboost]{xgb.train} for more details. Can be set to \emph{auto}
 in order to let the function choose the better model regarding the
 output variable.}

\item{verbose}{\strong{Logical}. Verbose parameter for grid search.}

\item{...}{Other parameters from \link[rBayesianOptimization]{BayesianOptimization}.}
}
\value{
The optimization results with the following fields:
\itemize{
 \item{\strong{param}}: the optimal set of parameters.
 \item{\strong{err}}: the error associated to the optimal parameter
 set.
 \item{\strong{results}}: the history of the results for the
 cross-validation with all the tested sets of parameters.
}
}
\description{
\code{xg_bo} use the bayesian optimization algorithm implemented in
 the \code{rBayesianOptimization} package in order
 to select the best set of parameters for an xgboost model.
}
\examples{
d <- xg_load_data(system.file("extdata", "titanic.csv", package = "ezXg"),
               inputs = c("Pclass", "Sex", "Age", "SibSp",
                          "Parch", "Fare", "Embarked"),
               output = "Survived",
               train.size = 0.8)
t <- xg_bo(d)

}
