% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgez.R
\name{xg_gs}
\alias{xg_gs}
\title{Grid search}
\usage{
xg_gs(data, eta = c(0.05, 0.1, 0.15, 0.2, 0.25, 0.3), gamma = c(0, 0.1,
  0.2, 0.3, 0.4, 0.5), max_depth = c(1, 3, 4, 5, 6, 8, 10, 12, 15),
  colsample_bytree = c(0.3, 0.4, 0.5, 0.7, 0.8, 0.9, 1),
  min_child_weight = c(1, 3, 5, 7), nrounds = 100, nthread = 2,
  cv = 5, seed = 1, verbose = TRUE, objective = "auto")
}
\arguments{
\item{data}{\strong{Object}. A data structure created by the call of the
\link[ezXg]{load_data} function.}

\item{eta}{\strong{Numeric Vectors}. Eta parameter list for grid search.
See \link[xgboost]{xgb.train} for more details.}

\item{gamma}{\strong{Numeric Vector}. Gamma parameter list for grid search.
See \link[xgboost]{xgb.train} for more details.}

\item{max_depth}{\strong{Numeric Vector}. Max_depth parameter list for grid
search. See \link[xgboost]{xgb.train} for more details.}

\item{colsample_bytree}{\strong{Numeric Vector}. Colsample_bytree parameter
list for grid search. See \link[xgboost]{xgb.train} for more details.}

\item{min_child_weight}{\strong{Numeric Vector}. Min_child_weight parameter
list for grid search. See \link[xgboost]{xgb.train} for more details.}

\item{nrounds}{\strong{Numeric}. Nrounds parameter for xgboost calibration.
See \link[xgboost]{xgb.train} for more details.}

\item{nthread}{\strong{Numeric}. Nthread parameter for xgboost calibration.
See \link[xgboost]{xgb.train} for more details.}

\item{cv}{\strong{Numeric}. Number of folds in cross validation. Needs
to be more than 2.}

\item{seed}{\strong{Numeric}. Seed for computation reproducability.}

\item{verbose}{\strong{Logical}. Verbose parameter for grid search.}

\item{objective}{\strong{Character}. Objective function for the
optimization. . Eta parameter for xgboost calibration. See
 \link[xgboost]{xgb.train} for more details. Can be set to \emph{auto}
 in order to let the function choose the better model regarding the
 output variable.}
}
\value{
The optimization results with the following fields:
\itemize{
 \item{\strong{param}}: the optimal set of parameters.
 \item{\strong{err}}: the error associated to the optimal parameter
 set.
 \item{\strong{results}}: the history of the results for the
 cross-validation with all the tested sets of parameters.
}
}
\description{
\code{xg_gs} use coordinate descent
 (\link{https://en.wikipedia.org/wiki/Coordinate_descent}) in order
 to select the best set of parameters for an xgboost model. At the end
 of the coordinate descent algorithm, a full search on each of the
 individual parameter vectors is made in order to potentially improve
 the selection.
}
\examples{
temp.file <- tempfile()
data(iris)
fwrite(iris, temp.file, row.names = F)
d <- load_data(temp.file, output = "Species", train.size = 0.8)
t <- xg_gs(d)

}
