% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgez.R
\name{xg_predict}
\alias{xg_predict}
\title{Predict on new values}
\usage{
xg_predict(model, data)
}
\arguments{
\item{model}{\strong{Object} A model object created by the function
\link[ezXg]{xg_train}.}

\item{data}{\strong{data.frame}. A data.frame or data.table structure with
column names equal to the input names.}
}
\value{
The prediction with the following fields:
\itemize{
 \item{\strong{pred}}: the prediction for the model.
 \item{\strong{proba}}: (optionnal) the associated probabilities for the
 prediction in case of a classification
}
}
\description{
\code{xg_predict} use the previously trained xgboost model to perform
prediction on new data.
}
\examples{
d <- xg_load_data(system.file("extdata", "titanic.csv", package = "ezXg"),
               inputs = c("Pclass", "Sex", "Age", "SibSp",
                          "Parch", "Fare", "Embarked"),
               output = "Survived",
               train.size = 0.8)
md <- xg_train(d)
p <- xg_predict(md, d$data[train == FALSE])

}
