% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgez.R
\name{xg_predict}
\alias{xg_predict}
\title{Predict on new values}
\usage{
xg_predict(model, data)
}
\arguments{
\item{model}{\strong{Object} A model object created by the function
\link[ezXg]{xg_train}.}

\item{data}{\strong{data.frame}. A data.frame or data.table structure with
column names equal to the input names.}
}
\description{
\code{xg_predict} use the previously trained xgboost model to perform
prediction on new data.
}
\examples{
temp.file <- tempfile()
data(iris)
fwrite(iris, temp.file, row.names = FALSE)
d <- load_data(temp.file, output = "Species", train.size = 0.8)
md <- xg_train(d)
p <- xg_predict(md, iris)

}
